remove some casts to short
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1825749 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/testcases/org/apache/poi/xssf/usermodel/TestXSSFColor.java
@@ -35,7 +35,7 @@ public final class TestXSSFColor {
       XSSFWorkbook wb = XSSFTestDataSamples.openSampleWorkbook("48779.xlsx");
 
       // Check the CTColor is as expected
-      XSSFColor indexed = wb.getCellStyleAt((short)1).getFillBackgroundXSSFColor();
+      XSSFColor indexed = wb.getCellStyleAt(1).getFillBackgroundXSSFColor();
       assertEquals(true, indexed.getCTColor().isSetIndexed());
       assertEquals(64, indexed.getCTColor().getIndexed());
       assertEquals(false, indexed.getCTColor().isSetRgb());
